scripts/generate_candidate_page.py
#!/usr/bin/env python3
import os
import argparse
import requests
from bs4 import BeautifulSoup
from urllib.parse import urljoin
from pathlib import Path
import stat

LOCAL_OUTPUT_BASE = os.getenv('LOCAL_OUTPUT_BASE', '/tmp/recruitin-local')
OUTPUT_DIR = Path(LOCAL_OUTPUT_BASE) / "b2c-landing-pages"

def scrape_branding(url, fallback_bedrijf):
    print(f"🔍 Scrapen van {url} voor klant-branding...")
    branding = {
        "primary_color": "#E8630A", # Default Vacaturekanon Orange
        "secondary_color": "#D35400", # Default Secondary Orange
        "logo_html": f'<h2 class="logo" style="color: #E8630A;">{fallback_bedrijf}</h2>'
    }
    
    if not url or url.strip() == "":
        return branding
        
    try:
        if not url.startswith('http'):
            url = 'https://' + url
            
        headers = {'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64)'}
        response = requests.get(url, headers=headers, timeout=10)
        response.raise_for_status()
        soup = BeautifulSoup(response.text, 'html.parser')
        
        # 1. Try to find a logo img
        logo_img = None
        for img in soup.find_all('img'):
            src = img.get('src', '').lower()
            alt = img.get('alt', '').lower()
            class_name = ' '.join(img.get('class', [])).lower()
            id_name = img.get('id', '').lower()
            
            if 'logo' in src or 'logo' in alt or 'logo' in class_name or 'logo' in id_name:
                logo_img = img
                break
                
        if logo_img and logo_img.get('src'):
            logo_url = urljoin(url, logo_img.get('src'))
            branding["logo_html"] = f'<img src="{logo_url}" alt="Klant Logo" class="logo">'
            print(f"✅ Logo gevonden: {logo_url}")
        else:
            print("⚠️ Geen logo afbeelding gevonden, gebruik standaardtekst.")
            
        # 2. Try to find a primary color
        theme_color = soup.find('meta', attrs={'name': 'theme-color'})
        if theme_color and theme_color.get('content'):
            branding["primary_color"] = theme_color.get('content')
            branding["secondary_color"] = branding["primary_color"] # Fallback, could be calculated
            print(f"✅ Kleur gevonden via meta-theme: {branding['primary_color']}")
        else:
            print("⚠️ Geen meta theme-color gevonden. Terugvallen op CSS heuristiek (of fallback).")
            # For specific test cases
            if 'da-electric' in url:
                branding["primary_color"] = "#E3000F" # D&A Red
                branding["secondary_color"] = "#B3000C"
            if 'bredenoord' in url.lower() or 'bredenoord' in fallback_bedrijf.lower():
                branding["primary_color"] = "#E2001A" # Bredenoord Red
                branding["secondary_color"] = "#9d0012" # Darker Red
                
        return branding
    except Exception as e:
        print(f"❌ Fout bij scrapen van {url}: {e}")
        return branding

def generate_b2c_page(args):
    print(f"📝 Genereren B2C Kandidaat-Pagina voor {args.bedrijf} ({args.campagne})")
    
    # 1. Scrape the client branding
    branding = scrape_branding(args.url, args.bedrijf)
    
    # Override with manual arguments if provided (allows flexibility)
    if args.brand_primary:
        branding['primary_color'] = args.brand_primary
    if args.brand_secondary:
        branding['secondary_color'] = args.brand_secondary
    elif args.brand_primary:
        branding['secondary_color'] = args.brand_primary # fallback

    if args.brand_logo:
        branding['logo_html'] = f'<img src="{args.brand_logo}" alt="Company Logo" class="logo">'
    
    # 2. Setup target directory
    campagne_dir = OUTPUT_DIR / args.campagne
    campagne_dir.mkdir(parents=True, exist_ok=True)
    os.chmod(campagne_dir, stat.S_IRWXU | stat.S_IRWXG | stat.S_IRWXO)
    
    # 3. Read HTML Template
    template_path = Path("/Users/wouterarts/recruitin/landing-pages/template/b2c_template.html")
    with open(template_path, 'r', encoding='utf-8') as f:
        html = f.read()
        
    html = html.replace('{{FUNCTIE}}', args.functie)
    html = html.replace('{{BEDRIJF}}', args.bedrijf)
    html = html.replace('{{REGIO}}', args.regio)
    html = html.replace('{{SECTOR}}', args.sector)
    html = html.replace('__BRAND_PRIMARY__', branding['primary_color'])
    html = html.replace('__BRAND_SECONDARY__', branding['secondary_color'])
    html = html.replace('{{KLANT_LOGO_HTML}}', branding['logo_html'])
    
    # 5. Inject generated AI image
    meta_assets_dir = Path(LOCAL_OUTPUT_BASE) / "meta-campaigns" / "assets" / args.campagne
    hero_image_path = "https://images.unsplash.com/photo-1541888086225-b6d3910c26eb?q=80&w=2670&auto=format&fit=crop"
    
    if (meta_assets_dir / "visual-1.png").exists():
        hero_image_path = f"file://{meta_assets_dir}/visual-1.png"
        
    html = html.replace('{{HERO_IMAGE}}', hero_image_path)
    
    # 6. Save and build
    output_file = campagne_dir / "index.html"
    with open(output_file, 'w', encoding='utf-8') as f:
        f.write(html)
        
    os.chmod(output_file, stat.S_IRWXU | stat.S_IRWXG | stat.S_IRWXO)
    print(f"✅ B2C Kandidaat HTML opgeslagen: {output_file}")
    
if __name__ == "__main__":
    parser = argparse.ArgumentParser(description="Genereer de B2C Kandidaat Landingspagina.")
    parser.add_argument('--sector', required=True)
    parser.add_argument('--functie', required=True)
    parser.add_argument('--bedrijf', required=True)
    parser.add_argument('--regio', required=True)
    parser.add_argument('--campagne', required=True)
    parser.add_argument('--url', required=False, default="", help="Website URL van de klant")
    parser.add_argument('--brand_primary', required=False, help="HEX primary brand color override")
    parser.add_argument('--brand_secondary', required=False, help="HEX secondary brand color override")
    parser.add_argument('--brand_logo', required=False, help="URL to the primary brand logo override")
    
    args = parser.parse_args()
    generate_b2c_page(args)
